src/scripts/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
idna==3.10
minio==7.2.16
numpy==2.3.3
paho-mqtt==2.1.0
pandas==2.3.2
pyarrow==21.0.0
pycparser==2.23
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
